Vue's composition API doesn't need inject to work like hooks :tada: - Separated fearch filters into their own file, will come in handy later - Line-clamping is such a bless :sparkles: - Fixed Card URLs opening the API reference, not the HTML repository
