+++ package.json
@@ -37,6 +37,7 @@
     "@nuxtjs/eslint-module": "^3.0.2",
     "@nuxtjs/tailwindcss": "^3.4.2",
     "@tailwindcss/forms": "^0.2.1",
+    "@tailwindcss/line-clamp": "^0.1.0",
     "@tailwindcss/postcss7-compat": "npm:@tailwindcss/postcss7-compat",
     "autoprefixer": "^9",
     "babel-eslint": "^10.1.0",
