+++ tsconfig.json
@@ -26,8 +26,7 @@
     "types": [
       "@types/node",
       "@nuxt/types",
-      "@nuxt/http",
-      "@vercel/node"
+      "@nuxt/http"
     ]
   },
   "exclude": [
